Move record code into separate class
